+++ examples/ci-circleci-orbs/scripts/greet.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+# A clean script for shellcheck to approve of.
+name="${1:-world}"
+echo "hello, ${name}"